+++ templates/express/config/passport.js
@@ -12,17 +12,42 @@ var mongoose = require('mongoose'),
 /**
  * Passport configuration
  */
-module.exports = function() {
-  passport.serializeUser(function(user, done) {
-    done(null, user.id);
+passport.serializeUser(function(user, done) {
+  done(null, user.id);
+});
+passport.deserializeUser(function(id, done) {
+  User.findOne({
+    _id: id
+  }, '-salt -hashedPassword', function(err, user) { // don't ever give out the password or salt
+    done(err, user);
   });
-  passport.deserializeUser(function(id, done) {
+});
+
+// add other strategies for more authentication flexibility
+passport.use(new LocalStrategy({
+    usernameField: 'email',
+    passwordField: 'password' // this is the virtual field on the model
+  },
+  function(email, password, done) {
     User.findOne({
-      _id: id
-    }, '-salt -hashedPassword', function(err, user) { // don't ever give out the password or salt
-      done(err, user);
+      email: email
+    }, function(err, user) {
+      if (err) return done(err);
+
+      if (!user) {
+        return done(null, false, {
+          message: 'This email is not registered.'
+        });
+      }
+      if (!user.authenticate(password)) {
+        return done(null, false, {
+          message: 'This password is not correct.'
+        });
+      }
+      return done(null, user);
     });
-  });
+  }
+));
 
   // add other strategies for more authentication flexibility
   passport.use(new LocalStrategy({
@@ -147,3 +172,5 @@ module.exports = function() {
   ));<% } %>
 
 };
+
+module.exports = passport;