+++ templates/express/routes.js
@@ -63,6 +63,11 @@ module.exports = function(app) {
       }), users.authCallback);
   <% } %>
 
+  // All undefined api routes should return a 404
+  app.get('/api/*', function(req, res) {
+    res.send(404);
+  });
+
   // All other routes to use Angular routing in app/scripts/app.js
   app.get('/partials/*', index.partials);
   app.get('/*',<% if(mongoPassportUser) { %> middleware.setUserCookie,<% } %> index.index);